Convert tab panel in entry editor to JavaFX
