+++ build.gradle
@@ -117,6 +117,7 @@ dependencies {
     compile 'net.corda:jfx:0.11.1'
     compile 'org.fxmisc.flowless:flowless:0.5.2'
     compile 'de.jensd:fontawesomefx-materialdesignfont:1.7.22-4'
+    compile 'org.fxmisc.richtext:richtextfx:0.7-M5'
 
 
     compile 'commons-logging:commons-logging:1.2'
