+++ src/main/java/org/jabref/gui/exporter/SaveDatabaseAction.java
@@ -129,9 +129,6 @@ public class SaveDatabaseAction extends AbstractWorker {
         }
 
         try {
-            // Make sure the current edit is stored
-            panel.storeCurrentEdit();
-
             // If set in preferences, generate missing BibTeX keys
             panel.autoGenerateKeysBeforeSaving();
